Moved recipe summary into its own component, created recipe details component, updated routing, and created temporary simulation of database
