+++ src/app/app.module.ts
@@ -4,11 +4,15 @@ import { FormsModule } from '@angular/forms'
 import { HttpClientModule } from '@angular/common/http'
 
 import { AppRoutingModule } from './app-routing.module'
-import { AppComponent } from './app.component'
+import { AppComponent } from './app.component';
+import { RecipeDetailsComponent } from './recipe-details/recipe-details.component';
+import { RecipesSummaryComponent } from './recipes-summary/recipes-summary.component'
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    RecipeDetailsComponent,
+    RecipesSummaryComponent
   ],
   imports: [
     BrowserModule,
