+++ src/app/core/services/data-service.service.ts
@@ -5,9 +5,23 @@ import { HttpClient } from '@angular/common/http'
   providedIn: 'root'
 })
 export class DataService {
+  recipes: any[] = []  // TODO: delete once db is implemented
+
   constructor(private http: HttpClient) { }
 
   getScrapedRecipe(recipeUrl: string) {
     return this.http.get('/api/recipe/' + recipeUrl)
+  }
+
+  // TODO: delete once db is implemented
+  storeRecipe(title: string, recipe: any): void {
+    this.recipes.push({
+      title: title,
+      recipe: recipe
+    })
+  }
+
+  getRecipe(title: string): void {
+    return this.recipes.find(r => r.title === title)
   }
 }
