+++ src/app/recipe-details/recipe-details.component.html
@@ -1,0 +1,2 @@
+<p>{{title}}</p>
+{{recipeData}}
